+++ src/js/core/scrollspy.js
@@ -70,11 +70,9 @@ export default function (UIkit) {
 
                     var index = this.elements.length === 1 ? 1 : 0;
 
-                    this.elements.each((_, el) => {
+                    this.elements.each((i, el) => {
 
-                        var $el = $(el);
-
-                        var data = el._scrollspy;
+                        var $el = $(el), data = el._scrollspy, cls = data.toggles[i] || data.toggles[0];
 
                         if (data.show) {
 
@@ -84,7 +82,7 @@ export default function (UIkit) {
 
                                     $el.css('visibility', '')
                                         .addClass(this.inViewClass)
-                                        .toggleClass(data.toggles[0])
+                                        .toggleClass(cls)
                                         .trigger('inview');
 
                                     data.inview = true;
@@ -104,16 +102,15 @@ export default function (UIkit) {
                                 }
 
                                 $el.removeClass(this.inViewClass)
-                                    .toggleClass(data.toggles[0])
+                                    .toggleClass(cls)
                                     .css('visibility', this.hidden ? 'hidden' : '')
                                     .trigger('outview');
 
                                 data.inview = false;
+
                             }
 
                         }
-
-                        data.toggles.reverse();
 
                     });